Змінив перевірку знаменника x,y,z та додав перевірку на результат функції NaN

+++ lab08/prj/main/my_lib.h
@@ -1,6 +1,6 @@
 #ifndef MY_LIB_H_INCLUDED
 #define MY_LIB_H_INCLUDED
-double s_calculation(double x, double y, double z);
+float s_calculation(float x, float y, float z);
 
 
 #endif // MY_LIB_H_INCLUDED
